tested, logged, improved
added some new APIs in controllers, configuring email messages

+++ src/main/java/ru/neostudy/loanConveyorProject/deal/entity/Credit.java
@@ -6,12 +6,14 @@ import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.Type;
 import org.hibernate.annotations.TypeDef;
+import org.springframework.stereotype.Component;
 import ru.neostudy.loanConveyorProject.conveyor.dto.PaymentScheduleElement;
 import ru.neostudy.loanConveyorProject.deal.enums.CreditStatus;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
 import java.util.List;
+
 
 @Getter
 @Setter
@@ -23,7 +25,7 @@ public class Credit {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY, generator="credit_id_seq")
     @Column(name = "credit_id", nullable = false)
-    private Integer creditId;
+    private Long creditId;
 
     @Column(name = "amount")
     private BigDecimal amount;
