+++ src/main/java/ru/neostudy/loanConveyorProject/deal/entity/Passport.java
@@ -16,7 +16,7 @@ public class Passport {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY, generator="passport_id_seq")
     @Column(name = "passport_id", nullable = false)
-    private Integer passportId;
+    private Long passportId;
 
     @Column(name = "series")
     private String series;
